% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.pars.cal.tab.suna.R
\name{def.pars.cal.tab.suna}
\alias{def.pars.cal.tab.suna}
\title{Parse SUNA cal table using calibration filename}
\usage{
def.pars.cal.tab.suna(
  calFilename = NULL,
  calTableName = "CVALTABLEA1",
  log = NULL
)
}
\arguments{
\item{calFilename}{The filename assiciated with the desired calibration table [character]}

\item{calTableName}{The calibration table name, defaults to "CVALTABLEA1" [character]}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A data frame of \cr
\code{wavelength} units = nanometer. \cr
\code{transmittance} units = unitless. \cr
}
\description{
Definition function. Given a calibration filename, the code parses the calibration table
if it exists. Some older SUNA files do not have a calibration table. As of the development
of this function no other sensors, besides the SUNA, have a calibraiton table.
}
\examples{
# TBD
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org}
}
\keyword{currently}
\keyword{none}
