% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.pars.data.suna.R
\name{def.pars.data.suna}
\alias{def.pars.data.suna}
\title{Parse SUNA data using data pasted in from def.wq.abs.corr.R}
\usage{
def.pars.data.suna(sunaBurst = NULL, log = NULL)
}
\arguments{
\item{sunaBurst}{The filename assiciated with the desired calibration table [dataframe]}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A vector of unitless transmittance values
}
\description{
Definition function. Given a burst of SUNA data, the code parses the spectrum_channels
and returns a vector of averaged transimittance intensities.
}
\examples{
# TBD
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org}
}
\keyword{currently}
\keyword{none}
